chore issue 708: Tooltip is one word
