+++ src/components/index.scss
@@ -49,5 +49,5 @@
 @import 'tabs/index';
 @import 'title/index';
 @import 'toast/index';
-@import 'tool_tip/index';
+@import 'tooltip/index';
 @import 'text/index';
