--- src/components/tool_tip/_index.scss
@@ -1,1 +1,0 @@
-@import 'tool_tip';
