--- src/components/tool_tip/index.js
@@ -1,7 +1,0 @@
-export {
-  EuiToolTip,
-} from './tool_tip';
-
-export {
-  EuiIconTip,
-} from './icon_tip';
