+++ src/components/tooltip/_index.scss
@@ -1,0 +1,1 @@
+@import 'tooltip';
