+++ src/components/tooltip/_tool_tip.scss
@@ -1,0 +1,113 @@
+.EuiTooltip {
+  @include euiBottomShadow;
+  @include euiFontSizeS();
+
+  position: absolute; /* 1 */
+  border-radius: $euiBorderRadius;
+  padding: $euiSizeM;
+  background-color: tintOrShade($euiColorFullShade, 25%, 90%);
+  color: $euiColorGhost;
+  max-width: 256px;
+  opacity: 0;
+  animation: EuiTooltipTop $euiAnimSpeedSlow ease-out $euiAnimSpeedNormal forwards;
+  z-index: $euiZLevel9;
+
+  &::before {
+    content: "";
+    position: absolute;
+    bottom: -$euiSize/2;
+    left: 50%;
+    transform: translateX(-50%) rotateZ(45deg);
+    transform-origin: center;
+    background-color: tintOrShade($euiColorFullShade, 25%, 90%);
+    width: $euiSize;
+    height: $euiSize;
+  }
+
+  // Positions the arrow and animates in from the same side.
+  &.EuiTooltip--right {
+    animation-name: EuiTooltipRight;
+
+    &:before {
+      bottom: 50%;
+      left: -$euiSize/2;
+      transform: translateY(50%) rotateZ(45deg);
+    }
+  }
+
+  &.EuiTooltip--bottom {
+    animation-name: EuiTooltipBottom;
+
+    &:before {
+      bottom: auto;
+      top: -$euiSize/2;
+    }
+  }
+
+  &.EuiTooltip--left {
+    animation-name: EuiTooltipLeft;
+
+    &:before {
+      bottom: 50%;
+      left: auto;
+      right: -$euiSize/2;
+      transform: translateY(50%) rotateZ(45deg);
+    }
+  }
+
+  .EuiTooltip__title {
+    font-weight: $euiFontWeightBold;
+    border-bottom: solid 1px tintOrShade($euiColorFullShade, 35%, 80%);
+    padding-bottom: $euiSizeXS;
+    margin-bottom: $euiSizeXS;
+  }
+}
+
+.EuiTooltipAnchor {
+  display: inline-block;
+}
+
+// Keyframes to animate in the tooltip.
+@keyframes EuiTooltipTop {
+  0% {
+    opacity: 0;
+    transform: translateY(-$euiSize);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@keyframes EuiTooltipBottom {
+  0% {
+    opacity: 0;
+    transform: translateY($euiSize);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@keyframes EuiTooltipLeft {
+  0% {
+    opacity: 0;
+    transform: translateX(-$euiSize);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@keyframes EuiTooltipRight {
+  0% {
+    opacity: 0;
+    transform: translateX($euiSize);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
