+++ src/components/tooltip/index.js
@@ -1,0 +1,7 @@
+export {
+  EuiTooltip,
+} from './tooltip';
+
+export {
+  EuiIconTip,
+} from './icon_tip';
