+++ src/components/tooltip/tooltip.test.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import { render } from 'enzyme';
+import { requiredProps } from '../../test';
+
+import { EuiTooltip } from './tooltip';
+
+describe('EuiTooltip', () => {
+  test('is rendered', () => {
+    const component = render(
+      <EuiTooltip title="title" id="id" content="content" {...requiredProps}>
+        <button>Trigger</button>
+      </EuiTooltip>
+    );
+
+    expect(component)
+      .toMatchSnapshot();
+  });
+});
